requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cloudinary==1.41.0
Django==4.1.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
frozenlist==1.4.1
idna==3.10
mongoengine==0.27.0
multidict==6.1.0
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
pymongo==3.12.3
six==1.16.0
sqlparse==0.2.4
tzdata==2024.2
urllib3==2.2.3
yarl==1.11.1
